refactor: coverImage 패키지 분할
